--- hawthorn-core/src/main/java/com/wunhwan/hawthorn/core/context/TargetContext.java
@@ -1,20 +1,0 @@
-package com.wunhwan.hawthorn.core.context;
-
-import com.wunhwan.hawthorn.core.metadata.TargetMetadata;
-import io.rsocket.core.RSocketClient;
-
-/**
- * todo...
- *
- * @author wunhwantseng@gmail.com
- * @since todo...
- **/
-public interface TargetContext {
-
-    Class<?> type();
-
-    TargetMetadata targetMetadata();
-
-    RSocketClient socketClient();
-
-}
