+++ hawthorn-core/src/main/java/com/wunhwan/hawthorn/core/metadata/TransferMetadata.java
@@ -9,7 +9,9 @@ package com.wunhwan.hawthorn.core.metadata;
 public class TransferMetadata {
 
     private String url;
+
     private String protocol;
+
     private String agreementType;
 
 }
