+++ hawthorn-transport/build.gradle.kts
@@ -5,6 +5,9 @@ dependencies {
     // core
     implementation("${rootProject.extra.get("rsocket-transport-netty")}")
 
+    // util
+    implementation("${rootProject.extra.get("commons-lang3")}")
+
     // test
     testImplementation("${rootProject.extra.get("jupiter-engine")}")
     testImplementation("${rootProject.extra.get("jupiter-api")}")
